LayoutTests/fast/css-grid-layout/grid-auto-flow-get-set-expected.txt

Test that setting and getting -webkit-grid-auto-flow works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting -webkit-grid-auto-flow set through CSS
PASS window.getComputedStyle(gridAutoFlowColumnSparse, '').getPropertyValue('-webkit-grid-auto-flow') is 'column'
PASS window.getComputedStyle(gridAutoFlowRowSparse, '').getPropertyValue('-webkit-grid-auto-flow') is 'row'
PASS window.getComputedStyle(gridAutoFlowDense, '').getPropertyValue('-webkit-grid-auto-flow') is 'row dense'
PASS window.getComputedStyle(gridAutoFlowColumnDense, '').getPropertyValue('-webkit-grid-auto-flow') is 'column dense'
PASS window.getComputedStyle(gridAutoFlowRowDense, '').getPropertyValue('-webkit-grid-auto-flow') is 'row dense'
PASS window.getComputedStyle(gridAutoFlowDenseColumn, '').getPropertyValue('-webkit-grid-auto-flow') is 'column dense'
PASS window.getComputedStyle(gridAutoFlowDenseRow, '').getPropertyValue('-webkit-grid-auto-flow') is 'row dense'
PASS window.getComputedStyle(gridAutoFlowInherit, '').getPropertyValue('-webkit-grid-auto-flow') is 'column'
PASS window.getComputedStyle(gridAutoFlowNoInherit, '').getPropertyValue('-webkit-grid-auto-flow') is 'row'

Test getting -webkit-grid-auto-flow bad values set through CSS
PASS window.getComputedStyle(gridAutoFlowNone, '').getPropertyValue('-webkit-grid-auto-flow') is 'row'
PASS window.getComputedStyle(gridAutoFlowColumns, '').getPropertyValue('-webkit-grid-auto-flow') is 'row'
PASS window.getComputedStyle(gridAutoFlowRows, '').getPropertyValue('-webkit-grid-auto-flow') is 'row'
PASS window.getComputedStyle(gridAutoFlowColumnFoo, '').getPropertyValue('-webkit-grid-auto-flow') is 'row'
PASS window.getComputedStyle(gridAutoFlowColumnColumn, '').getPropertyValue('-webkit-grid-auto-flow') is 'row'
PASS window.getComputedStyle(gridAutoFlowDenseRowRow, '').getPropertyValue('-webkit-grid-auto-flow') is 'row'

Test the initial value
PASS element.style.webkitGridAutoFlow is ''
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-flow') is 'row'

Test getting and setting -webkit-grid-auto-flow through JS
PASS element.style.webkitGridAutoFlow is 'column'
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-flow') is 'column'
PASS element.style.webkitGridAutoFlow is 'column dense'
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-flow') is 'column dense'
PASS element.style.webkitGridAutoFlow is 'row dense'
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-flow') is 'row dense'
PASS element.style.webkitGridAutoFlow is 'row dense'
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-flow') is 'row dense'
PASS element.style.webkitGridAutoFlow is 'column dense'
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-flow') is 'column dense'
PASS element.style.webkitGridAutoFlow is 'row dense'
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-flow') is 'row dense'
PASS element.style.webkitGridAutoFlow is 'row'
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-flow') is 'row'

Test getting and setting bad values for -webkit-grid-auto-flow through JS
PASS element.style.webkitGridAutoFlow is ''
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-flow') is 'row'
PASS element.style.webkitGridAutoFlow is ''
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-flow') is 'row'
PASS element.style.webkitGridAutoFlow is ''
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-flow') is 'row'

Test setting -webkit-grid-auto-flow to 'initial' through JS
PASS element.style.webkitGridAutoFlow is 'initial'
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-flow') is 'row'
PASS successfullyParsed is true

TEST COMPLETE
